File Uploaded

+++ controllers/HomeController/homeController.js
@@ -1,30 +1,36 @@
-const Tweet = require("../../models/TweetModel/tweetModel");
-const Like = require("../../models/LikeModel/likeModel");
+const Tweet = require('../../models/TweetModel/tweetModel');
+const Like = require('../../models/LikeModel/likeModel');
 
 const dashboard = async (req, res) => {
+  try {
+    const allTweets = await Tweet.find({})
+      .populate('user', 'username avatar')
+      .sort({ createdAt: -1 });
+    const likes = await Like.find({ user: req.user.id }).select('tweet');
+    const likedTweetIds = likes.map((like) => like.tweet.toString());
 
-  const allTweets = await Tweet.find({}).populate('user', 'username avatar');
-
-  const likes = await Like.find({ user: req.user.id }).select("tweet");
-
-  const likedTweetIds = likes.map(like => like.tweet.toString());
-  
-  res.render("index", {
-    loginSuc: req.flash('loginSuc')[0],
-    loginSucIcon: req.flash("loginSucIcon")[0],
-    loginSucMsg: req.flash("loginSucMsg")[0],
-    addTweetSuc: req.flash("addTweetSuc")[0],
-    addTweetIcon: req.flash("addTweetIcon")[0],
-    addTweetSucMsg: req.flash("addTweetSucMsg")[0],
-    tweets : allTweets,
-    likedTweetIds,
-  });
+    res.render('index', {
+      loginSuc: req.flash('loginSuc')[0],
+      loginSucIcon: req.flash('loginSucIcon')[0],
+      loginSucMsg: req.flash('loginSucMsg')[0],
+      addTweetSuc: req.flash('addTweetSuc')[0],
+      addTweetIcon: req.flash('addTweetIcon')[0],
+      addTweetSucMsg: req.flash('addTweetSucMsg')[0],
+      tweets: allTweets,
+      likedTweetIds,
+      user: req.user,
+      userId: req.user.id, // Pass userId explicitly
+    });
+  } catch (err) {
+    console.error('Error fetching dashboard:', err);
+    res.status(500).send('Internal Server Error');
+  }
 };
 
 const addTweet = async (req, res) => {
   const { desc } = req.body;
   const path = req.file ? req.file.path : '';
-  
+
   try {
     const tweet = await Tweet.create({
       desc,
@@ -34,59 +40,56 @@ const addTweet = async (req, res) => {
       retweetCount: 0,
     });
 
-    if (tweet) {
-      console.log("Tweet Added.", tweet);
-      req.flash("addTweetSuc", "You Just Tweeted!");
-      req.flash("addTweetIcon", "🐦");
-      req.flash(
-        "addTweetSucMsg",
-        "Your thoughts are flying high in the Twitterverse! 🌍"
-      );
-      res.redirect("/twitter-clone/home/");
-    } else {
-      console.log("Error : Tweet Not Added.");
-    }
+    req.flash('addTweetSuc', 'You Just Tweeted!');
+    req.flash('addTweetIcon', '🐦');
+    req.flash('addTweetSucMsg', 'Your thoughts are flying high in the Twitterverse! 🌍');
+    res.redirect('/twitter-clone/home/');
   } catch (err) {
-    console.log("Internal Server Error :- ", err);
+    console.error('Internal Server Error:', err);
+    res.status(500).send('Internal Server Error');
   }
 };
 
 const handleLikes = async (req, res) => {
-
   const { tweetId, action } = req.body;
 
-  try{
+  if (!tweetId || !['increment', 'decrement'].includes(action)) {
+    return res.status(400).json({ success: false, message: 'Invalid tweetId or action' });
+  }
+
+  try {
     const tweet = await Tweet.findById(tweetId);
-    const socket = req.app.get('socket');
+    if (!tweet) {
+      return res.status(404).json({ success: false, message: 'Tweet not found' });
+    }
 
-    if(action === 'increment'){
+    const io = req.app.get('io');
+    let likeCount = tweet.likeCount;
 
+    if (action === 'increment') {
       const existingLike = await Like.findOne({ user: req.user.id, tweet: tweetId });
-
       if (!existingLike) {
         await Like.create({ user: req.user.id, tweet: tweetId });
         tweet.likeCount++;
+        likeCount++;
       }
-
-    }else{
-
+    } else {
       const deletedLike = await Like.findOneAndDelete({ user: req.user.id, tweet: tweetId });
-      
       if (deletedLike) {
         tweet.likeCount = Math.max(0, tweet.likeCount - 1);
+        likeCount = tweet.likeCount;
       }
     }
 
     await tweet.save();
 
-    socket.emit('likeUpdate', {  tweetId, userId : req.user.id, likeCount : tweet.likeCount, action });
+    io.emit('likeUpdate', { tweetId, userId: req.user.id, likeCount, action });
 
-    console.log('like added.');
-
-  }catch(err){
-    console.log('Internal Server Error :', err);
+    res.json({ success: true, likeCount });
+  } catch (err) {
+    console.error('Internal Server Error:', err);
+    res.status(500).json({ success: false, message: 'Internal Server Error' });
   }
-
-}
+};
 
 module.exports = { dashboard, addTweet, handleLikes };
